[config] Move PeerNetworkId to be with NetworkId
Remove the upstream config file, as it no longer exists
