+++ mempool/src/shared_mempool/peer_manager.rs
@@ -10,7 +10,10 @@ use crate::{
         types::{notify_subscribers, SharedMempool, SharedMempoolNotification},
     },
 };
-use diem_config::config::{MempoolConfig, PeerNetworkId, PeerRole, RoleType};
+use diem_config::{
+    config::{MempoolConfig, PeerRole, RoleType},
+    network_id::PeerNetworkId,
+};
 use diem_infallible::Mutex;
 use diem_logger::prelude::*;
 use diem_types::transaction::SignedTransaction;
